Re-add parsing errors
These are compatible, the added test passes before the introduction of JXKit in the previous commit.

This unfortunately can't be implemented by catching the JS error object as JXKit hides this, presenting just the string representation. This could be improved with upstream contributions to JXKit, it looks as if it would fit in their API, but that's out of scope for now.

This could also be implemented in a nicer way with regular expressions, but to do so would require bumping the minimum deployment target much further forward. Instead we just defensively parse the lines, falling back to a generic error message if we don't get the right structure.